Move InputValidationLengthConstraint to the common module
- move InputValidationLengthConstraint to the common module from the forum
- update forum
- update content working group
