+++ runtime-modules/common/Cargo.toml
@@ -10,6 +10,8 @@ std = [
 	'sr-primitives/std',
 	'srml-support/std',
 	'system/std',
+	'codec/std',
+	'serde'
 ]
 
 
@@ -30,3 +32,14 @@ default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-system'
 rev = 'c37bb08535c49a12320af7facfd555ce05cce2e8'
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.0.0'
+
+[dependencies.serde]
+features = ['derive']
+optional = true
+version = '1.0.101'
