+++ runtime-modules/common/src/lib.rs
@@ -1,5 +1,6 @@
 // Ensure we're `no_std` when compiling for Wasm.
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod constraints;
 pub mod currency;
 pub mod origin_validator;
