+++ runtime-modules/forum/Cargo.toml
@@ -61,7 +61,7 @@ git = 'https://github.com/paritytech/substrate.git'
 package = 'srml-balances'
 rev = 'c37bb08535c49a12320af7facfd555ce05cce2e8'
 
-[dev-dependencies.common]
+[dependencies.common]
 default_features = false
 package = 'substrate-common-module'
 path = '../common'
@@ -90,4 +90,5 @@ std = [
   	'balances/std',
 	'timestamp/std',
 	'bureaucracy/std',
+	'common/std',
 ]
