+++ runtime-modules/proposals/engine/Cargo.toml
@@ -20,7 +20,6 @@ std = [
     'sr-primitives/std',
     'membership/std',
     'common/std',
-
 ]
